export less function in public and move def. to cpp
